+++ frontend/src/router/index.js
@@ -1,5 +1,7 @@
 import { createRouter, createWebHistory } from 'vue-router'
-import HomeView from '../views/HomeView.vue'
+import HomeView from '@/views/HomeView.vue'
+import PetsView from '@/views/PetsView.vue'
+import PetView from '@/views/PetView.vue'
 
 const router = createRouter({
   history: createWebHistory(import.meta.env.BASE_URL),
@@ -12,10 +14,21 @@ const router = createRouter({
     {
       path: '/about',
       name: 'about',
+      component: HomeView
+    },
+    {
+      path: '/pets',
+      name: 'Pets',
       // route level code-splitting
       // this generates a separate chunk (About.[hash].js) for this route
       // which is lazy-loaded when the route is visited.
-      component: () => import('../views/AboutView.vue')
+      component: PetsView
+    },
+    {
+      path: '/pet/:id',
+      name: 'Pet',
+      component: PetView,
+      props: true
     }
   ]
 })