+++ CHANGELOG.md
@@ -5,7 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.3.0] - 2023-01-10
+## [0.3.1] - 2023-01-10
+### Added
+- Added delete methods
+- Autopopulate age based on DOB
 ### Changed
 - Return IDs of related models
 - Update VueJS homepage with example code